hardcoded impl bug fix

+++ repository/src/main/java/by/itacademy/sologub/StudentRepoHardcodedImpl.java
@@ -11,10 +11,10 @@ import static by.itacademy.sologub.constants.Constants.*;
 
 @Slf4j
 public class StudentRepoHardcodedImpl implements StudentRepo {
-    static int CURRENT_MAX_STUDENT_ID = 100;
+    static int CURRENT_MAX_STUDENT_ID = 6789;
     private static StudentRepoHardcodedImpl instance;
     private final CredentialRepo credentialRepo;
-    private final Map<Credential, Student> students;
+    private final Map<Integer, Student> students;
 
     private StudentRepoHardcodedImpl(CredentialRepo credentialRepo) {
         this.credentialRepo = credentialRepo;
@@ -41,9 +41,13 @@ public class StudentRepoHardcodedImpl implements StudentRepo {
     public Student getStudentIfExistsOrGetSpecialValue(String login) {
         Credential cr = credentialRepo.getCredentialIfExistsOrGetSpecialValue(login);
         if (LOGIN_NOT_EXISTS == cr) {
+            log.info("Логина не существует, возвращаем специальный обьект {}", LOGIN_NOT_EXISTS);
             return STUDENT_NOT_EXISTS;
         } else {
-            return students.get(cr);
+            log.info("Логин существует, патаемся вернуть студента");
+            return students.values().stream()
+                    .filter(s -> s.getCredential().getLogin().equals(login))
+                    .findAny().orElse(STUDENT_NOT_EXISTS);
         }
     }
 
@@ -51,11 +55,17 @@ public class StudentRepoHardcodedImpl implements StudentRepo {
     public Student getStudentIfExistsOrGetSpecialValue(String login, String password) {
         Credential cr = credentialRepo.getCredentialIfExistsOrGetSpecialValue(login, password);
         if (LOGIN_NOT_EXISTS == cr) {
+            log.info("Логина не существует, возвращаем специальный обьект {}", LOGIN_NOT_EXISTS);
             return STUDENT_NOT_EXISTS;
         } else if (PASSWORD_WRONG == cr) {
+            log.info("Прароль не верен, возвращаем специальный обьект {}", STUDENT_PASSWORD_WRONG);
             return STUDENT_PASSWORD_WRONG;
         } else {
-            return students.get(cr);
+            log.info("Логин и пароль соответствуют, патаемся вернуть студента");
+            return students.values().stream()
+                    .filter(s -> s.getCredential().getLogin().equals(login) &&
+                            s.getCredential().getPassword().equals(password))
+                    .findAny().orElse(STUDENT_NOT_EXISTS);
         }
     }
 
@@ -63,19 +73,21 @@ public class StudentRepoHardcodedImpl implements StudentRepo {
     public boolean putStudentIfNotExists(Student student) {
         Credential cr = credentialRepo.getCredentialIfExistsOrGetSpecialValue(student.getCredential().getLogin());
         if (LOGIN_NOT_EXISTS == cr) {
+            log.info("логина не существует, значит можно добавить нового студента");
             String login = student.getCredential().getLogin();
             String password = student.getCredential().getPassword();
 
             credentialRepo.putCredentialIfNotExists(login, password);
-            Credential studentCr = credentialRepo.getCredentialIfExistsOrGetSpecialValue(login, password);
+            cr = credentialRepo.getCredentialIfExistsOrGetSpecialValue(login, password);
 
-            if (isExistsAndPasswordRight(studentCr)) {
+            if (isExistsAndPasswordRight(cr)) {
                 student.setId(CURRENT_MAX_STUDENT_ID++);
-                student.setCredential(studentCr);
-                students.put(studentCr, student);
+                student.setCredential(cr);
+                students.put(student.getId(), student);
+                log.info("Удалось положить Credential, добавяем нового студента {}", student);
                 return true;
             } else {
-                return false;
+                log.info("Не удалось положить Credential, нельзя добавить нового студента");
             }
         }
         return false;
@@ -83,24 +95,23 @@ public class StudentRepoHardcodedImpl implements StudentRepo {
 
     @Override
     public boolean changeStudentParametersIfExists(Student newS) {
-        Credential oldCred = credentialRepo.getCredentialIfExistsOrGetSpecialValue(newS.getCredential().getLogin());
-        Student oldS = students.get(oldCred);
+        Student oldS = students.get(newS.getId());
         log.debug("Пытаемся менять параметры Student и Credential на новые");
 
         if (oldS != null && STUDENT_NOT_EXISTS != oldS) {
             log.debug("Меняем параметры Student и Credential на новые");
 
+            oldS.setFirstname(newS.getFirstname());
+            oldS.setLastname(newS.getLastname());
+            oldS.setPatronymic(newS.getPatronymic());
+            oldS.setDateOfBirth(newS.getDateOfBirth());
             boolean isChanged = credentialRepo.changeCredentialIfExists(newS.getCredential().getLogin(), newS.getCredential().getPassword());
             if (isChanged) {
-                oldS.setFirstname(newS.getFirstname());
-                oldS.setLastname(newS.getLastname());
-                oldS.setPatronymic(newS.getPatronymic());
-                oldS.setDateOfBirth(newS.getDateOfBirth());
                 log.info("Учётная запись изменена, можно менять параметры студента");
-                return true;
             } else {
-                log.info("Учётная запись не изменена, параметры студента тоже не будут изменены");
+                log.info("Учётная запись не изменена, но параметры студента будут изменены");
             }
+            return true;
         } else {
             log.debug("Нельзя менять параметры Student или Credential если их не существует в репозитории");
         }
@@ -113,13 +124,21 @@ public class StudentRepoHardcodedImpl implements StudentRepo {
         log.debug("Пытаемся проверить учетные данные обьекта Student {} в репозитории", cr);
         if (LOGIN_NOT_EXISTS == cr) {
             log.info("Объекта учётных данных нет в базе. Нечего удалять");
-            return false;
         } else {
-            students.remove(cr);
-            credentialRepo.deleteCredentialIfExists(login);
-            log.info("Объекты Student и Credentials удалены из репозиториев");
-            return true;
+            Student s = students.values().stream()
+                    .filter(student -> student.getCredential().getLogin().equals(cr.getLogin()))
+                    .findAny().orElse(STUDENT_NOT_EXISTS);
+            if (STUDENT_NOT_EXISTS != s) {
+                s.setCredential(null);// help GC
+                students.remove(s.getId());
+                credentialRepo.deleteCredentialIfExists(login);
+                log.info("Объекты Student и Credentials удалены из репозиториев");
+                return true;
+            } else {
+                log.info("Объекта студентов нет в базе. Нечего удалять");
+            }
         }
+        return false;
     }
 
     @Override
